requirements.txt
aiohttp
fastapi
loguru
passlib
PyJWT
tortoise-orm[asyncpg]
uvicorn
python-dotenv
ujson
apscheduler
python-dateutil
pydantic_settings